requirements.txt
# ========================================
# 本番環境用 依存パッケージ定義ファイル (requirements.txt)
# ========================================
# 概要: このプログラムを実行するために必要な外部ライブラリのリスト
#       pipコマンドでこのファイルを読み込むと、必要なライブラリが自動インストールされる
#
# なぜこのファイルが必要？:
#   - プロジェクトで使用する外部ライブラリを明確に記録
#   - 開発環境を他の人と共有する際に、同じバージョンのライブラリをインストール可能
#   - 本番環境へのデプロイ時に必要なライブラリを自動インストール
#   - セキュリティ脆弱性のあるバージョンを避けるため、バージョンを固定
#
# 使い方:
#   pip install -r requirements.txt
#
# Python バージョン要件: 3.11以上
# ========================================

# ====================
# Discord関連
# ====================
discord.py==2.3.2  # Discord Botを作成するためのライブラリ（VCメンバー取得、サーバー接続用）

# ====================
# Google Drive/Sheets関連
# ====================
gspread==6.1.0  # Google Sheetsを操作するためのライブラリ（スプレッドシートの読み書き用）
google-auth==2.29.0  # Google APIの認証を行うライブラリ（サービスアカウント認証用）
google-auth-oauthlib==1.2.0  # OAuth2.0認証をサポートするライブラリ（より高度な認証方式用）
google-api-python-client==2.134.0  # Google Drive APIを使用するためのライブラリ（CSVファイル管理用）

# ====================
# Slack関連（オプション）
# ====================
slack-sdk==3.27.1  # Slack APIを使用するためのライブラリ（通知機能用、使用は任意）

# ====================
# 環境変数管理
# ====================
python-dotenv==1.0.1  # .envファイルから環境変数を読み込むライブラリ（秘密情報の管理用）

# ====================
# タイムゾーン処理
# ====================
pytz==2024.1  # タイムゾーンを扱うライブラリ（日本時間への変換用）

# ====================
# 非同期処理サポート
# ====================
aiohttp==3.9.5  # 非同期HTTPクライアント/サーバーフレームワーク（Discord.pyが内部で使用）
asyncio-mqtt==0.16.2  # MQTT通信の非同期処理用（将来の拡張用、現在は未使用）

# ====================
# ユーティリティ
# ====================
typing-extensions==4.11.0  # Python型ヒントの拡張機能（型チェック機能の強化用）
jpholiday==0.1.10  # 日本の祝日判定ライブラリ（営業日ベースの連続ログイン日数計算用）
loguru==0.7.2  # 高機能なロギングライブラリ（詳細なログ出力用）